requirements.txt
# requirements.txt (tương thích cả Pi & Windows nếu cài đúng môi trường)
kivy==2.3.1
kivymd==1.2.0
pillow==10.4.0
requests==2.32.3
urllib3==2.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
filetype==1.2.0
docutils==0.21.2
python-dotenv==1.1.0
